+++ src/accessories/contact.js
@@ -1,14 +1,10 @@
 import Logger from '../helper/logger.js';
-import moment from 'moment';
-
-const timeout = (ms) => new Promise((res) => setTimeout(res, ms));
 
 export default class ContactAccessory {
-  constructor(api, accessory, accessories, tado, deviceHandler, FakeGatoHistoryService) {
+  constructor(api, accessory, accessories, tado, deviceHandler) {
     this.api = api;
     this.accessory = accessory;
     this.accessories = accessories;
-    this.FakeGatoHistoryService = FakeGatoHistoryService;
 
     this.deviceHandler = deviceHandler;
     this.tado = tado;
@@ -72,7 +68,7 @@ export default class ContactAccessory {
       service.addCharacteristic(this.api.hap.Characteristic.ClosedDuration);
 
     service.getCharacteristic(this.api.hap.Characteristic.ResetTotal).onSet((value) => {
-      Logger.info(value + ': Resetting FakeGato..', this.accessory.displayName);
+      Logger.info(value + ': Resetting history counters..', this.accessory.displayName);
 
       const now = Math.round(new Date().valueOf() / 1000);
       const epoch = Math.round(new Date('2001-01-01T00:00:00Z').valueOf() / 1000);
@@ -86,33 +82,11 @@ export default class ContactAccessory {
         .updateValue(this.accessory.context.timesOpened);
     });
 
-    this.historyService = this.FakeGatoHistoryService ? new this.FakeGatoHistoryService('door', this.accessory, {
-      storage: 'fs',
-      path: this.api.user.storagePath(),
-      disableTimer: true,
-    }) : undefined;
-
-    await timeout(250); //wait for historyService to load
-
     service
       .getCharacteristic(this.api.hap.Characteristic.ContactSensorState)
       .on(
         'change',
-        this.deviceHandler.changedStates.bind(this, this.accessory, this.historyService, this.accessory.displayName)
+        this.deviceHandler.changedStates.bind(this, this.accessory, this.accessory.displayName)
       );
-
-    if (this.FakeGatoHistoryService && !this.refreshHistoryHandlerRegistered) {
-      this.deviceHandler.refreshHistoryHandlers.push(() => this.refreshHistory(service));
-      this.refreshHistoryHandlerRegistered = true;
-    }
-  }
-
-  refreshHistory(service) {
-    let state = service.getCharacteristic(this.api.hap.Characteristic.ContactSensorState).value;
-
-    if (this.historyService) this.historyService.addEntry({
-      time: moment().unix(),
-      status: state ? 1 : 0,
-    });
   }
 }