+++ src/accessories/humidity.js
@@ -1,14 +1,10 @@
 import Logger from '../helper/logger.js';
-import moment from 'moment';
-
-const timeout = (ms) => new Promise((res) => setTimeout(res, ms));
 
 export default class HumidityAccessory {
-  constructor(api, accessory, accessories, tado, deviceHandler, FakeGatoHistoryService) {
+  constructor(api, accessory, accessories, tado, deviceHandler) {
     this.api = api;
     this.accessory = accessory;
     this.accessories = accessories;
-    this.FakeGatoHistoryService = FakeGatoHistoryService;
 
     this.deviceHandler = deviceHandler;
     this.tado = tado;
@@ -25,7 +21,7 @@ export default class HumidityAccessory {
 
     if (!service) {
       Logger.info('Adding HumiditySensor service', this.accessory.displayName);
-      service = this.accessory.addService(
+      this.accessory.addService(
         this.api.hap.Service.HumiditySensor,
         this.accessory.displayName,
         this.accessory.context.config.subtype
@@ -50,35 +46,5 @@ export default class HumidityAccessory {
       }
     }
 
-    this.historyService = this.FakeGatoHistoryService ? new this.FakeGatoHistoryService('room', this.accessory, {
-      storage: 'fs',
-      path: this.api.user.storagePath(),
-      disableTimer: true,
-    }) : undefined;
-
-    await timeout(250); //wait for historyService to load
-
-    service
-      .getCharacteristic(this.api.hap.Characteristic.CurrentRelativeHumidity)
-      .on(
-        'change',
-        this.deviceHandler.changedStates.bind(this, this.accessory, this.historyService, this.accessory.displayName)
-      );
-
-    if (this.FakeGatoHistoryService && !this.refreshHistoryHandlerRegistered) {
-      this.deviceHandler.refreshHistoryHandlers.push(() => this.refreshHistory(service));
-      this.refreshHistoryHandlerRegistered = true;
-    }
-  }
-
-  refreshHistory(service) {
-    let state = service.getCharacteristic(this.api.hap.Characteristic.CurrentRelativeHumidity).value;
-
-    if (this.historyService) this.historyService.addEntry({
-      time: moment().unix(),
-      temp: 0,
-      humidity: state,
-      ppm: 0,
-    });
   }
 }