+++ src/accessories/motion.js
@@ -1,14 +1,10 @@
 import Logger from '../helper/logger.js';
-import moment from 'moment';
-
-const timeout = (ms) => new Promise((res) => setTimeout(res, ms));
 
 export default class MotionAccessory {
-  constructor(api, accessory, accessories, tado, deviceHandler, FakeGatoHistoryService) {
+  constructor(api, accessory, accessories, tado, deviceHandler) {
     this.api = api;
     this.accessory = accessory;
     this.accessories = accessories;
-    this.FakeGatoHistoryService = FakeGatoHistoryService;
 
     this.deviceHandler = deviceHandler;
     this.tado = tado;
@@ -41,33 +37,11 @@ export default class MotionAccessory {
     if (!service.testCharacteristic(this.api.hap.Characteristic.LastActivation))
       service.addCharacteristic(this.api.hap.Characteristic.LastActivation);
 
-    this.historyService = this.FakeGatoHistoryService ? new this.FakeGatoHistoryService('motion', this.accessory, {
-      storage: 'fs',
-      path: this.api.user.storagePath(),
-      disableTimer: true,
-    }) : undefined;
-
-    await timeout(250); //wait for historyService to load
-
     service
       .getCharacteristic(this.api.hap.Characteristic.MotionDetected)
       .on(
         'change',
-        this.deviceHandler.changedStates.bind(this, this.accessory, this.historyService, this.accessory.displayName)
+        this.deviceHandler.changedStates.bind(this, this.accessory, this.accessory.displayName)
       );
-
-    if (this.FakeGatoHistoryService && !this.refreshHistoryHandlerRegistered) {
-      this.deviceHandler.refreshHistoryHandlers.push(() => this.refreshHistory(service));
-      this.refreshHistoryHandlerRegistered = true;
-    }
-  }
-
-  refreshHistory(service) {
-    let state = service.getCharacteristic(this.api.hap.Characteristic.MotionDetected).value;
-
-    if (this.historyService) this.historyService.addEntry({
-      time: moment().unix(),
-      status: state ? 1 : 0,
-    });
   }
 }