+++ src/helper/telegram.js
@@ -1,5 +1,4 @@
 import Logger from './logger.js';
-import FormData from 'form-data';
 
 export default class Telegram {
   constructor(options, messages) {
@@ -7,13 +6,7 @@ export default class Telegram {
     this.chatID = options.chatID;
     this.messages = messages;
 
-    this.request = {
-      protocol: 'https:',
-      host: 'api.telegram.org',
-      port: 443,
-      method: 'POST',
-      path: '/bot' + this.token + '/sendMessage',
-    };
+    this.url = `https://api.telegram.org/bot${this.token}/sendMessage`;
   }
 
   send(target, dest, replacer, additional) {
@@ -27,27 +20,26 @@ export default class Telegram {
 
       const form = new FormData();
 
-      this.request.headers = form.getHeaders();
       form.append('chat_id', this.chatID);
       form.append('parse_mode', 'Markdown');
       form.append('text', message);
 
       Logger.debug('Telegram: Sending Message: ' + message);
 
-      form.submit(this.request, (err, res) => {
-        if (err) {
+      fetch(this.url, { method: 'POST', body: form })
+        .then((res) => {
+          if (res.status < 200 || res.status > 200) {
+            Logger.error('A response error occured during sending telegram message!');
+            Logger.error({
+              code: res.status,
+              message: res.statusText,
+            });
+          }
+        })
+        .catch((err) => {
           Logger.error('An error occured during sending telegram message!');
           Logger.error(err);
-        }
-
-        if (res.statusCode < 200 || res.statusCode > 200) {
-          Logger.error('A response error occured during sending telegram message!');
-          Logger.error({
-            code: res.statusCode,
-            message: res.statusMessage,
-          });
-        }
-      });
+        });
     } else {
       Logger.debug('Telegram: Skip sending, no message defined for ' + target);
     }